ocamlrep_marshal fwd compat with ocaml-5.0
Summary:
this hack enables `ocamlrep_marshal` to marshal in the correct format regardless of whether that's ocaml-4.14.0 marshal format or the ocaml-5.0.0 marshal format.

the hack enables us to move forward with testing hack against ocaml-5.0.0 without any source level changes. when we've migrated and are on 5.0.0 exclusively the hack can be removed.

Reviewed By: vsiles

Differential Revision: D43177579

fbshipit-source-id: 66b2596438c60024cfd4ae327235033fc284d5cf