Include more rules
Summary:
Previously we grabbed the rules from prelude, which meant that those we didn't export (because we override with stubs) got lost. Instead use rules as the source of truth.

At the same time, move to using load_symbols, which is much better than native, and allows unwinding some documentation hacks.

Reviewed By: stepancheg

Differential Revision: D44786906

fbshipit-source-id: 430d2f96bf93e37016ba1cb9fbd284b4883c57c4