Extract CompileContext to its own module
Summary:
I am interested in adding 2 new caches to this type in D44933138, and using those from a module that is not build.bzl.

Starlark requires that the modules are structured as a DAG i.e. cycles when performing `load` are not allowed.

In order for more of the code to be able to use the same CompileContext for state that is global to the entire analysis of a single target, it is more convenient for it to get its own module.

Reviewed By: zertosh

Differential Revision: D44932822

fbshipit-source-id: 24abb5eebfd3411e2b9079218272012cc092f6ca