Add new apple_focused_debugging rule
Summary:
We want to be able to filter the debug data we materialize for a build, based upon whether the user is utilizing focused debugging.

In order to do that, we need to surface to buck2, which targets users want in the build. To facilitate that, add a `apple_focused_debugging` rule, which contains include/exclude of the build target patterns and regular expressions the user has used to generate their workspace.

This information will be used later on to
1. filter the debug info downloaded for each binary
2. inform (up the stack) the focused debugging binary scrubber which paths to scrub (D44937947)
3. determine if we're already downloading the archives needed for the link and should link locally or remotely

Reviewed By: milend

Differential Revision: D44941272

fbshipit-source-id: 43b8209b7ca0edeed7238fa2ab4dcdc957a42376