Scrub binary and filter debug data before bundling
Summary:
The an `apple_bundle` has `focused_debugging` set, we want to only bundle the scrubbed binary and filtered debug data.

So when we get the binary, make sure it is scrubbed. At the same time, generate the filtered dsym and debuginfo.

The test plan covers manual testing of the integration.

I'll followup with a more comprehensive test that we can run e2e.

Reviewed By: blackm00n

Differential Revision: D45130219

fbshipit-source-id: aba842de830a6d28aa471a3f9f6fd2b44a0035ea