Tighten API for SwiftDependencyInfo
Summary:
Within an `apple_library`, there is only one `SwiftDependencyInfo` object (verified by reading through the code, as well as checking if we ever have more than one for large apps like Instagram).

While it might be convenient to utilize an array and empty array in the codebase, it becomes difficult to reason about the code, and even causes unnecessary iteration.

For example, the primary code point in which dependency info for swift is formed takes the info from deps and combines it:
https://www.internalfb.com/code/fbsource/[a43509e337bd27fc61941e5373b920e968e1ae92]/fbcode/buck2/prelude/apple/swift/swift_compilation.bzl?lines=570%2C573

But then end up iterating over an item of one:
https://www.internalfb.com/code/fbsource/[a43509e337bd27fc61941e5373b920e968e1ae92]/fbcode/buck2/prelude/apple/apple_library.bzl?lines=155%2C209%2C288-293

Tightening this API allows us to reason about the code better in the future.

Reviewed By: maxovtsin

Differential Revision: D45569016

fbshipit-source-id: db859d75ab6d364a43d620d7c7e7269df15c2aff