Added `linker_flags` feature to link group definition
Summary:
For any `cxx_library` you can define linker flags that will be propagated to the linker argsfile. We need to have the same for the link groups.
The real scenario that we need that is:
We have `cxx_unittest` macro that currently receives `scrs` with test files. To improve cacheability of the graph, we are gonna split that like:
```
cxx_unittest(
   name = "my_test",
   deps = [":my_test-lib"],
)

cxx_library(
   name = "my_test-lib",
   srs = ["test_source_file.cpp"],
)
```
The thing is, linker will ignore adding `my_test-lib.so` into NEEDS definition of the result binary, because "my_test" binary don't use any references from "my_test-lib". To address that issue, we can define library like
```
cxx_library(
   name = "my_test-lib",
   srs = ["test_source_file.cpp"],
   linker_flags = ["-Wl,--no-as-needed"],
)
```
enforcing the linkage if library is linked dynamically.

The thing is, we need to define "my_test-lib" as the link groups to make it being linked dynamically in `dev-lg` mode to improve cacheability.
And we need to have same explicit linker flag propagation that will enforce shared object to be added to executable binary.

Reviewed By: christycylee

Differential Revision: D45183540

fbshipit-source-id: a6f8003f27a1bb444affc65fbe2080be572efefc