Use target label on the AndroidPackageableInfo for module seeds
Summary:
When constructing our Voltron target graph, we always use the label from the `AndroidPackageableInfo`. This means that aliases are not included in the graph (since they don't actually contribute anything), but it does mean that if we use the target label of an alias, we won't find a node in the graph.

Therefore, this fixes a bug where we were using the target label from the buck2 node, rather than from the provider attached to that node.

Reviewed By: navidqar

Differential Revision: D45729299

fbshipit-source-id: 37f206632cff6cdab3a77c3cf4f43c56949e7b5c