Update apple_macro_layer_set_bool_override_attrs_from_config to take a structured record
Summary:
In the next diff, we need to update `apple_macro_layer_set_bool_override_attrs_from_config` to handle more than just True or False values, and so that when the config is false, to not set values.

Specifically utilize a record that defines
- the attribute name
- buck config section
- buck config key
- values to match when the override is positive
- value to set when positive
- value to set when negative
- whether to skip setting when negative

As such, update the method and existing calls.

Reviewed By: blackm00n

Differential Revision: D45746136

fbshipit-source-id: 4d172db06f2f7e46b2513deae858ca92f92e3f82