Replace link_locally_override with link_execution_preference
Summary:
In D44446714, Apple's `link_locally_override` was hoisted to be on the Cxx rules. However, it's currently [unused](https://www.internalfb.com/code/search?q=repo%3Afbsource%20link_locally_override).

This makes it easy for us to replace this with `link_execution_preference`, following a similar approach to the previous diff.

For Apple rules, it is controlled by config.

For Cxx rules, it is set on the target itself (stripped for buck1).

Reviewed By: krallin

Differential Revision: D45706138

fbshipit-source-id: e2ecaa8cbf20e8deddda064f6b89a1da764f79b2