Replace `_link_libraries_locally_override` with link_execution_preference
Summary:
The `_link_libraries_locally_override` attribute was added the D40140013 stack to allow a buck config to set whether the apple_library should link locally or not.

With `link_execution_preference` we can now represent both that and whether we want to link remotely.

This allows us to solely use `link_execution_preference` in replacement of `_link_libraries_locally_override`.

Reviewed By: blackm00n

Differential Revision: D45706137

fbshipit-source-id: 3d8f809e133b0e13e05bba37c7bc34440128eaf1