Only test path basename for scale suffix
Summary:
When testing for asset suffixes, like foo@2x.png, we should only be looking for these characters in the basename of the src path.

For example, if the path is in a scoped dependency package (e.g., .../foo/bar/qux@2x.png), we should only strip the scale suffix from the path basename (qux@2x.png).

Reviewed By: IanChilds

Differential Revision: D45974667

fbshipit-source-id: 5a6507b62b6826e489970a709a16e324daae5ad2