Add do_not_create_inner_jar flag for java binary buck rule
Summary:
Adds do_not_create_inner_jar to prevent the creation of a jar within a jar if native libs are present. This diff allows the creation of an inner.jar or a single flat jar.

Our project needs this to export jars that can be imported by sbt projects.

Reviewed By: IanChilds

Differential Revision: D46074319

fbshipit-source-id: 97e2b32b2a01c29eb9b9b66e37d9a7168122b3d0