Use worker rule for javacd and kotlincd
Summary:
Calling `WorkerInfo` within `prepare_cd_exe` effectively creates a new worker for each action, having it be specified by a shared rule ensures that the worker itself is shared since the identity of the worker is essentially the starlark pointer value of the `WorkerInfo`.

This way in a given build there should be a single javacd worker and a single kotlincd worker.

Reviewed By: IanChilds

Differential Revision: D46222807

fbshipit-source-id: f8d2628671bf1ba9b9fb544899a59e4ef00532be