Sub-target for cxx_library to produce bitcode bundles (merged .bc file)
Summary: This is a change to `cxx_library` whereby a subtarget of ["bitcode"] exists for libraries.  This is just the object files for the library at hand, and only for static.  Shared will be supported as well, but will have the concept of both direct objects, as well as the sum of objects plus static library bundles.

Differential Revision: D45632239

fbshipit-source-id: 295a0b39eb7370d422d0cd1df8c4440cba8bc06e