Dedupe CxxCompileCommandOutputForCompDb into CxxCompileCommandOutput
Summary:
In D37477974, `CxxCompileCommandOutputForCompDb` was introduced to allow setting compile commands that were for the compilation db only:

https://www.internalfb.com/code/fbsource/[455092982907ae1fbd7d1e2ca8117edb6ff7dae9]/fbcode/buck2/prelude/cxx/compile.bzl?lines=128-134

This is so that for header only libraries, we can add them to the comp db:

https://www.internalfb.com/code/fbsource/[455092982907ae1fbd7d1e2ca8117edb6ff7dae9]/fbcode/buck2/prelude/cxx/compile.bzl?lines=275-284

We don't need to include two `CxxCompileCommandOutput` in order to achieve this. We can simply add a another field that is for compilation db only and then dedupe/delete the `CxxCompileCommandOutputForCompDb` record.

This simplifies the code and makes it easier to add additional arguments to `CxxCompileCommandOutput` without unnecessarily duping additional code.

Lastly, added meaningful defaults to `CxxCompileCommandOutput` to make it easier to create with less fields.

Differential Revision: D46325905

fbshipit-source-id: 39b44bc066ed15fcac04b5de25072c5287774888