buck2 - add qpld_dotslash to build action
Summary:
We need QPLd service to be able to send the QPL data from RE and we need to deploy qpld dotslash with the build action as [suggested by Remote Execution team](https://fb.workplace.com/groups/1042353022615812/permalink/2254272498090519/).

We need to qpld_dotslash to be available in buck2 builds (optionally in buck1) in RE and not in local (although this diff doesn't distinguish that) and the absence of qpld_dotslash should not fail builds.

In this diff, I pass qpld_dotslash to Kotlin compiler for buck2

**Next Steps:**
1. Connect `shouldUseCompilationTracer` to a build_rule param so that compilation tracer doesn't work for every build in RE
2.  [Optional] Make "passing qpld_dotslash" only for RE

Reviewed By: IanChilds

Differential Revision: D46108006

fbshipit-source-id: a4d1545b91f6bc40ee82332d41dcf1342a219918