Use import_lib when provided to prebuilt_cxx_library
Summary:
On Windows, it's illegal to link against a dll directly.
That's why we have the import_lib attribute. Use that if its non-None as the actual linkable. This way we still propagate the runtime library as a runtime dep. But, only do this shenanigan if we're targeting Windows as that's the only platform that uses this behavior, as otherwise users have to put the import lib attribute into a select that only provides it when *on* windows.

Reviewed By: jdonald, lmvasquezg

Differential Revision: D46187611

fbshipit-source-id: 0956c4e89089639e3256769943d77fbd1f8e4a22