Java_binary with native_libs and a main class should always have an inner jar
Summary: The main purpose for *not* having an inner jar is if the jar will be used as an exported dependency for a project that doesn't use buck. For performance reasons, its a bad practice to have a jar with native libs without an inner jar *if* you have a main class you are running.

Reviewed By: IanChilds

Differential Revision: D46366872

fbshipit-source-id: 441cad65e18e4f112832b55378fc3688b6f795ec