Symlink ghci binary
Summary:
## This stack
See summary of D46476480.

## This diff
Handle the ghc binary properly for the two cases:
1. a custom binary is provided in the `ghci_bin_dep` attribute. This is used in haxlsh.
2. use the default binary provided by the haskell toolchain. sigma/hhh uses this.

I can't use the buck2 `ctx.actions.symlink_file` function with the default binary provided by the toolchain because it's provided in a string format, and the method requires an artifact (see post Sergey left in the comment). I need to pass it as a target/dependency to use it, but I don't want to do this now bc it's not strictly necessary, so I'll follow up later.

Reviewed By: pepeiborra

Differential Revision: D46476563

fbshipit-source-id: c86aa311c4bee952597858c16620e7eb6d1078dc