Write start.ghci file
Summary:
## This stack
This stack is what I have so far for the native `haskell_ghci` buck2 rule implementation.
On D44449945 I published an initial RFC with what I had so far, but there were still a lot of things I didn't know.

Now, I think I'm much closer to getting to finishing, but there are still a few things I don't know that could be causing the errors that I'm getting...
I'll try to aggregate everything in an Help section...

My goal now is to
1. Generate scripts that are equivalent (e.g. only differences are in directory paths)
2. Generate all the necessary artifacts (e.g. SOs, scripts), which I'll compare using `tree -lv BUCK_OUT` of buck1 and buck2 output.
3. Run haxlsh and successfully load and run an endpoint.

I'm very close to #1. The only thing that's different is that I'm not creating an omnibus SO and I'm passing all the dependencies explicitly.

## This diff
As title.

Reviewed By: pepeiborra

Differential Revision: D46476480

fbshipit-source-id: aa9eea0bd0ed9b65b94156f27adda16fd69afe7e