Use CxxAdditionalArgsfileParams.argsfiles to augment relative and absolute argsfiles
Summary:
Now that we populate `CxxAdditionalArgsfileParams.argsfiles` in the Swift code (only consumer of this API), we can now use it to correctly augment `argsfile_by_ext` and `abs_argsfile_by_ext` with the respective argsfiles.

With this diff, ***now correctly populate `abs_argsfiles_by_ext` with the necessary swift argsfiles***.

This renders `CxxAdditionalArgsfileParams.argsfiles_deprecated` unusued and we will clean it up at the top of the stack.

Reviewed By: milend

Differential Revision: D46743446

fbshipit-source-id: ba9e402d90648364e87bdd20b4d921289b7d8d37