Detects failures from all types of CT hooks callbacks
Summary:
At the moment, cth_tpx would not see failures coming from certain CTH callbacks, so it could report a test as passing, while in reality it was supposed to fail. In a nuthsell, the problem is that to ensure cth_tpx sees all failures, it would need to be the the last callback to be executed in all post_* functions, but since execution order for "end_per" callbacks is reversed wrt "init_per" callbacks, this is simply impossible (see explanation in the cth_tpx_role.erl file).

So we solve this by having two instances of cth_tpx running, with highest (well, good effort, very high) priority and one with lowest, and assigning different roles to each. That is, each role handles only those callbacks for which it is in the ideal position (first or last to run). Again, see cth_tpx_role.erl for the details.

Reviewed By: TheGeorge

Differential Revision: D46764195

fbshipit-source-id: 4866cc4e9709d5af7a4b0e2c6dde4f4b450d4880