Update proc-macro2 from 1.0.63 to 1.0.64
Summary:
This fixes a bug related to bare carriage returns that caused D47318798 to fail in `third-party-rust-verification`. The `schemafy_core` crate contains a file that uses carriage returns which are not followed by newlines 🤮. When that file fails to parse, Reindeer falls back to using non-precise srcs, which produced the following difference to the generated output.

```
 diff --git a/third-party/rust/BUCK b/third-party/rust/BUCK
 --- a/third-party/rust/BUCK
+++ b/third-party/rust/BUCK
@@ -77047,6 +77047,7 @@
     name = "schemafy_core-0.5.1",
     srcs = [
         "vendor/schemafy_core-0.5.1/src/lib.rs",
+        "vendor/schemafy_core-0.5.1/src/main.rs",
         "vendor/schemafy_core-0.5.1/src/one_or_many.rs",
     ],
     crate = "schemafy_core",
```

Reviewed By: zertosh

Differential Revision: D47318928

fbshipit-source-id: 086b96393cc43371b724ac95e841990385576432