Use read_root_config in android rules
Summary:
We want to remove `read_config` one day, and only live with `read_root_config` and `read_package_value`.

It is probably not possible to remove all of them in the whole repository, but let's start somewhere.

This one should be safe, because android targets live in the root cell.

Reviewed By: IanChilds

Differential Revision: D47140221

fbshipit-source-id: 0e0fa01c6a1ca4b4bbcbbf3192078be9fd476c2b