Add pic_behavior to CxxToolchainInfo
Summary:
Within `buck2`'s prelude, we have three different `LinkStyle`'s: `shared`, `static` and `static_pic`.

`pic` behaves differently every platform:

- Windows: not recognized, `static_pic` is the same as `static`.
- Apple: `-fPIC` is enabled implicitly. `static_pic` and `static` are the same.
- Linux: `-fPIC` is respected.

This creates an issue in buck2: We might declare/ask for artifacts which are a mix of `static` + `static_pic` on Windows + Apple are creating identical artifacts/compile_cmds which create the exact same output.

To make this more correct we will introduce `pic_behavior` on the `CxxToolchainInfo` which will help us decide how to handle the `LinkStyle` on a per-platform basis:

- `PicBehavior("not_supported")`: akin to Windows, this means we don't recognize that compile flag on this platform.
- `PicBehavior("supported")`: akin to Linux, if you specify it, it's obeyed.
- `PicBehavior("always_enabled")`: akin to Apple, regardless of whether the flag is set, the value is applied.

Upstream I will add the logic to obey `pic_behavior` on the platform.

Reviewed By: milend

Differential Revision: D47197094

fbshipit-source-id: 7e19ab69a73f8b2a72b6ac4726f16ea75a48eb15