Allow remote execution for Pyre BXL scripts
Summary:
This diff updates Pyre's BXL scripts to allow remote execution, which includes moving common logic between `classic.bxl` and `code_navigation.bxl` for getting manifests into `merge.bxl`.

To do this we use `ctx.bxl_actions()` instead of `ctx.actions_factory` and supply the `exec_deps` parameter, which declares the merger target as a required dependency.

An earlier attempt at this caused S350428 because the merger target was not an execution dependency, and blocked the deprecation of `ctx.actions_factory`.

Reviewed By: zsol

Differential Revision: D47349741

fbshipit-source-id: e6e110efd0949782a2290fc0b116da1b3c6ba990