Revert D47382297: Multisect successfully blamed "D47382297: [buck2][prelude][cxx] Track C/C++ sources and expose via `[sources]` sub-target" for test or build failures
Summary:
This diff is reverting D47382297
D47382297: [buck2][prelude][cxx] Track C/C++ sources and expose via `[sources]` sub-target by andrewjcg has been identified to be causing the following test or build failures:

Tests affected:
- [buck2/tests/e2e/python:test_rules_with_compiled_buck2 - test_rules.py::test_omnibus_metadata[omnibus-roots]](https://www.internalfb.com/intern/test/281475068875740/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2452478
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: blackm00n

Differential Revision: D47491208

fbshipit-source-id: 56f6e3242009e4eb2b2d38f6d0c100b76314e188