Use same JavaCD+KotlinCD binaries for worker and non-worker actions
Summary:
Since we're no longer using the specified `main_class` (instead we pass it to ClassLoaderBootstrapper), and these binaries contains the same classes, we can just use a single copy of the binary.

The fixes the issue where we were still using the non-worker binary in the `BUCK_CLASSPATH` arg even when we were using workers (which did work fine, since the binaries contain the same paths, but just seems odd).

Reviewed By: christolliday

Differential Revision: D47510628

fbshipit-source-id: 667ee0bdc38152f6c2d360719f2a42b6e3f99a39