Export split debug handling (take 4)
Summary:
When using split-debug modes and LTO on macos or linux, the link step will emit
debug info objects that we need to track via external build info.

Export this directory in the `LinkedObject` struct, so that we can access it from
the caller.

Previous attempt caused some issues w/ Rust links using LTO+split-dwarf, which
have been cleaned up in this previous diffs in this stack:
https://fb.prod.workplace.com/groups/buck2users/posts/3462074467382243/

Reviewed By: athmasagar

Differential Revision: D47476866

fbshipit-source-id: 989d6401e2ed51544278f12cb0d2cbe712fca722