Scrub preprocessed java classes
Summary:
The command that users pass to preprocess `.jar` files is likely not deterministic, so we need to "scrub" the output jars to make sure that they are deterministic.

For each jar that is preprocessed, we create a new action where we run the zip scrubber on it.

(I was noticing that fb4a release builds were rebuilding a lot more than I expected, and I tracked it down to this issue, hence the fix).

Reviewed By: ianlevesque

Differential Revision: D47551016

fbshipit-source-id: e3ef28af7cab5dd4773501171ff5936236689c25