Replace support_pic with pic_behavior
Summary:
In D45645496 we added `support_pic` to avoid passing `-fPIC` to Windows compile commands.

This 'works' in that it avoid passing the flag, but didn't stop buck2 from producing duplicate `.pic.a` and `.a` artifacts, both of which would be compiled without `-fPIC`.

Since then we've added `PicBehavior` which will not only avoid the `-fPIC` flag, but, creating artifacts with duplicate contents. This isn't needed anymore.

Reviewed By: dtolnay

Differential Revision: D47268203

fbshipit-source-id: 5034e87875a34b8a116483272774a64f7e2c6d08