rust: don't emit empty "unused_extern_names" diagnostics
Summary:
Empty `unused_extern_names` is just noise. It's not like we emit
"message" lines when there are no rustc or clippy lints. This helps when
running over many targets since you're not filled with empty
`unused_extern_names` lines.

Moved `diag["level"] == "error"` to the `else` block so that it's
clearer in the code that we don't expect `level` for
`unused_extern_names` messages.

Reviewed By: dtolnay

Differential Revision: D47708228

fbshipit-source-id: 0721c2848a790a9bd881fbf2715c2112975af867