don't create explicit artifacts for framework logs
Summary:
The entire execution directory is already made available as archived artifact.

Most of the time these debug logs for the frameworks are completely uninteresting for developers.

This improves the dev experience by not linking the framework logs into the artifact directory. In addition, we also don't link .json files anymore as logs.

Reviewed By: jcpetruzza

Differential Revision: D47716467

fbshipit-source-id: 7ae52df8613269aa4c0226228ea8c4f3eae75f74