Don't pass DLL as linkable on windows
Summary:
On Windows, it's illegal to link against a dll directly. Right now if we add a dep that doesn't have an import lib the build errors even if the dep is not used/linked against. Let's allow the library to be created but not linked.

Fix for issue described in D47684657

Reviewed By: KapJI, milend

Differential Revision: D47716946

fbshipit-source-id: 241f65af88e415efdaed396f81cf8151e2594d8e