Support for non-asset prebuilt-native-libraries in modules
Summary:
This is the buck2 counterpart for the buck1 version that was added in D42826027 (we should try to add support for both at the same time in future, since this is quite a headache to track down).

Principle is the same - if we have native libraries that are not assets, then add them to the "module -> library" mapping that is produced by the `android_app_modularity` rule (this is then consumed later on).

Reviewed By: ianlevesque

Differential Revision: D47835090

fbshipit-source-id: 02075fc4cbb8d360e45054524fb05386106fff4c