Revert D47813466: Multisect successfully blamed "D47813466: [buck2][rust] Split DWARF support in Rust" for test or build failures
Summary:
This diff is reverting D47813466
D47813466: [buck2][rust] Split DWARF support in Rust by dtolnay has been identified to be causing the following test or build failures:

Tests affected:
- [scm/myles/daemon:daemon_rs-unittest - telemetry::tests::fill_daemon_query_argument_fields](https://www.internalfb.com/intern/test/562950054344658/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2595604
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: mdko

Differential Revision: D47865988

fbshipit-source-id: fe9bce8e5af62806c5e4b07ad99b0e9551d0dc32