Add support for FBReport asserts in buck2
Summary:
In buck2, FBReportWarning set up to break in the debugger if it occurs in "focused modules". In buck1, we use FocusedModules.txt in app bundle to get these focused modules.
In buck2,  since we use selective debugging now, we named this file as SelectedDebugPaths.txt.  In this diff, we set the file name in compile time using compiler flag.

Reviewed By: chatura-atapattu

Differential Revision: D47824036

fbshipit-source-id: 7e7cd640b118db17f29c8baa3036f0ab0adc2c26