Refactor cxx compile outputs
Summary:
This currently has a bunch of duplicated fields for pic and non-pic.
And then since those are optional each field is optional. With the change,
some code is deduplicated, but more importantly it's more clear what states this
can really be in (i.e. if we have pic outputs then we have all the related
fields).

Reviewed By: chatura-atapattu

Differential Revision: D47685101

fbshipit-source-id: 1a63e07a5666de0ece72d943767c7162d7e96011