Use weight=2 for javacd and kotlincd
Summary:
Benchmarking this has been somewhat inconclusive, since this mostly affects very slow benchmarks that have a wide variance in run time.

A weight of 5 seems more optimal on 80 core machines but setting it to 5 instead of using a concurrency limit of 16 regressed slow benchmarks on mybuild by ~50% because it cuts the effective concurrency limit from 16 down to 6.

Although this may not be optimal with 80 cores, the difference in performance there is less than it is on smaller machines, and we can optimize that later with a concurrency limit or by tuning GC args.

This doesn't seem make a big difference on 10 core macbooks either, and doesn't overload them, which makes sense, since this is roughly equivalent to the limits we already have in place for buck1.

Reviewed By: IanChilds

Differential Revision: D48045363

fbshipit-source-id: 76e199b284754fde71e69229545dc33896cde6ee