fix ocamlrep include path issues
Summary:
this has to have been broken for some time. it's unclear why it has gone undetected *update: actually it is clear and a fix for that is in D48102368*.

the issues fixed here apply to sandcastle and aren't as likely to exhibit for one reason or another on a devserver.

(1) ocamlopt.opt needs to be in path
(2) (a supercaml) ocamlopt.opt can return a standard library path location that doesn't exist

this diff ensures (1) and provides a workaround for (2).

Reviewed By: zertosh

Differential Revision:
D48102385

Privacy Context Container: L1124100

fbshipit-source-id: 06ee9bf79c174f7d3fdc3f9603c7925147f74e43