Add [dumpbin][headers] subtarget
Summary:
This diff adds a `[dumpbin][headers]` subtarget to any target with a pdb subtarget. It contains the output of `dumpbin /HEADERS <executable>`.
Eventually, this artifact will be used in install workflow for automatic symbol server upload. For now, just expose this in a subtarget so we can use it later for manual symbol upload.

(Note: this ignores all push blocking failures!)

Reviewed By: shayne-fletcher

Differential Revision: D48289437

fbshipit-source-id: 47bc682cf433ab89ac8043ad973b637483160066