Set a more compatible working directory
Summary:
Setting a `-working-directory` of `.` (to give all compilation actions the current working directory, whether running locally or on RE) results in the swift compiler injecting `././` into the dwarf data.

We need to set a working directory because our paths are relative and without it, none of the includes, etc can be found. So use `-working-directory=` to have the compiler use the current working directory without injecting an extra `./` into the dwarf data.

Reviewed By: blackm00n

Differential Revision: D48480135

fbshipit-source-id: 46e2f5f050485c8261005c23d1edd1954ca76ec8