Add debuginfo field to JavaClassToSourceMapInfo
Summary:
The classmap-parser tool that is used to generate debug information for DAP server consists of 2 parts:
1. Generating output for a set of java source files
2. Merging generated outputs into a single file
In buck2 we generate an action graph based on dependency graph for 1). This will be captured in "tset_debuginfo" field on the JavaClassToSourceMapInfo provider.
Additionally "debuginfo" field will contain a merged artifact.

Optinally [debuginfo] subtarget provides an access to "debuginfo" artifact as default output and can be used by non-BXL clients (and for testing)

Reviewed By: IanChilds

Differential Revision: D48468886

fbshipit-source-id: 438bc81367e47ad0261249a060ee6bf4ea1facda