Only force filter_resources to run locally when we have a post_filter_resources_cmd
Summary:
Running this action locally is really slow - I think it's because there's a lot of I/O contention, and it does a LOT of copying things around.

We shouldn't force this action to run locally unless it needs to, which is only the case if the label is set and we have a post_filter_resources_cmd.

Reviewed By: jiawei-lyu

Differential Revision: D48778181

fbshipit-source-id: cedb32a31bbdcf703a83c81af5432337ee319a75