Support passing --preferred-density to aapt2 link from android_binary and android_bundle
Summary:
It makes sense to support this flag, since it's a good (the best?) way to just get the densities we want in our builds.

For now (see next diffs), it will just be used to filter out unwanted densities in builds that will otherwise crash because they have too many entries to `zip`, but it could in future be used for filtering resources.

Reviewed By: jiawei-lyu

Differential Revision: D48780711

fbshipit-source-id: cbeb779baf88009a552295e39f0209ab36418f91