Remove panicking {To,From}OcamlRep impls for OsStr on wasm32
Summary:
I'm not sure that it makes sense to implement To/FromOcamlRep with a method that just raises a panic. We should be able to just decline to implement these traits at all for OsStr (and Path, etc) on wasm32 instead.

This could cause build breakage if we have users defining structs like:

```
#[derive(ToOcamlRep)]
struct MyStruct {
  path: PathBuf,
}
```

in crates which they build for the wasm32 target. But an attempt to make use of that ToOcamlRep impl would unconditionally panic at runtime anyway.

Reviewed By: shayne-fletcher

Differential Revision: D48842618

fbshipit-source-id: 840720a1b1a591640351f098fd2c8e2974ebcbf6