javacd_jar action cannot produce a class ABI if we have additional classes to merge
Summary: `javacd_jar` will produce a class ABI from only the srcs it is given, and won't include the additional classes to merge. So, don't produce one when we have additional classes to merge, and instead just rely on the class ABI generation that kicks in later on the merged output jar.

Reviewed By: jiawei-lyu

Differential Revision: D48949526

fbshipit-source-id: 375da0fcdd8d34ca5b9c8706ac10dad925b1ef65