Clarify use of "merge_link_infos"
Summary:
merge_link_infos was being used for two things:

1. create a MergeLinkInfo to propagate out of a target that doesn't add anything to it itself.
2. create a MergeLinkInfo that just holds data to be pulled out by get_link_args() later on and to be used for the target itself.

(2) is inefficient (creates a bunch of tsets to just through them away later)
and it makes the meaning of MergeLinkInfo itself confusing as it means that sometimes
it's a meaningful thing about how a target is added to a link and sometimes its just
a convenient data structure to pass around data. For example, for cxx_library,
it meant that we'd create two MergeLinkInfos, one of which would have special
handling for exported_deps, one that didn't and you'd need to carefully follow
which one is used where to understand that they are for two totally different
purposes.

This changes get_link_args to accept a list of MergeLinkInfo so there's no need
to merge the deps together and renames merge_link_info (and documents it) to make
it clearer that it should only be used for case (1).

Reviewed By: chatura-atapattu

Differential Revision: D47808803

fbshipit-source-id: 6d79343b525361ef853b2f36e91b73583350bf6e