add linker flags to linkablenode
Summary:
To support link strategies that look only at the LinkableGraph, we
need to have the linker flags args represented there. Sadly,
exported_linker_flags is handled implicitly already for most things because we
add it to the linkinfos for a pic_archive already, but some interesting
link strategies may still need to know it. So this just makes all four linker
flags fields available on LinkableNode.

Reviewed By: christolliday

Differential Revision: D49139112

fbshipit-source-id: caeb812f5ad23c69d1580df9f2461d5c47ce1552