add linkargs/shlib_deps to sharedlibrary and LinkedObject
Summary:
These are needed to better understand what produced the shared lib and
do interesting processing later on. The link_args will be used by android's
relinker to construct the same linker command but with an added version script
arg.

The shlib_deps we should actually move to later on actually using for determining
what shared libraries to bundle with top-level things (instead of right now where
we just literally take every shared lib in the graph or none of them).

Reviewed By: christolliday

Differential Revision: D49141120

fbshipit-source-id: fb19378eebf4781ef3becf7c01cc4eb19f3d2859