`export_exe` which is like `export_file` but accepts an arg
Summary:
Many times we want to take a group of files from the repo and
expose one (or more) of them as executable targets. Right now the only
way to do that is via a command_alias which uses a 'shim' launcher
like a .bat file or something and accepts an arg which points to the binary.
In buck2, we can write a rule which does this for us. This also removes
the need to use export_file + command_alias to expose a standalone binary
as a valid executable file.

Reviewed By: stepancheg

Differential Revision: D49079617

fbshipit-source-id: 236eb810607b925227c3d8450055126a5e6403f7