add support for native_library_merge_sequence
Summary:
In the rules, this is mostly about support for having an external
script produce the merge mapping. We serialize out the linkablegraph and the
apk module mapping.

This allows us then to implement the complex merge sequence logic all in an
isolated python script, that's significantly easier to understand and iterate
on.

This doesn't exactly match buck1, but that code is currently changing as well,
with this in python we can pretty easily update it to bring in any improvements
we have for buck1.

Reviewed By: mzlee

Differential Revision: D49141117

fbshipit-source-id: 867ad07ebcdf9523b2b2c4b78f6fa3b2b61d67ee