use CLANG_MODULE_CACHE_PATH for default module cache location
Summary:
A minor cleanup of the swift_exec wrapper. Rely on the CLANG_MODULE_CACHE_PATH env var to set the default module cache path for Swift when in an RE worker. The local value will point to the Darwin cache dir, which should work fine.

Ideally we would set this on the action itself, but env values seem to be escaped so we cannot reference the $TMPDIR env var.

Reviewed By: milend

Differential Revision: D49293646

fbshipit-source-id: eda79750a9488b9146153bc992e5da67563c9893