Forward `-warnings-as-errors` during explicit modules
Summary:
Reported in https://fb.workplace.com/groups/iosciqa/permalink/25194091246879486 and T165122233.

Buck2 using xbat was showing warnings as errors during the implicit module compilation, while Pika did not seem to do the same when using explicit modules.

Forward `-warning-as-errors` to the compilation of each module to replicate the behaviour of the implicit modules.

Reviewed By: rmaz

Differential Revision: D49751914

fbshipit-source-id: 23b4df254d541dc6e858d07d22b1a0b139329fd7