Replace filter-error and filter-ignore args from the Xcode argsfiles
Summary:
D48908378 will add these arguments to the argsfile:
```
"-filter-error=deprecated-declarations=is deprecated: first deprecated in iOS [2-5]"
-Xclang
"-filter-ignore=deprecated-declarations='stringWithCString:' is deprecated: first deprecated in iOS 2.0"
-Xclang
"-filter-ignore=deprecated-declarations='writeToFile:atomically:' is deprecated: first deprecated in iOS 2.0"
```

These are not compatible with Xcode indexing and need to be removed.

To do so, we're going to use the new performant API made available from core: `regex` and provide that to `replace_regex`. Original diff D49233548 was reverted (D49383884) due to non-performance.

Reviewed By: milend

Differential Revision: D50024531

fbshipit-source-id: ac7360f8ac8f7ce1656c7e5c0c0a2f9695227a82