Re-allow inplace-inplace to run remotely
Summary:
We noticed that inplace-inplace became local-only which means people don't get
cache hits for that, and it adds overhead on CAF builds of targets like ads_dper3.

With a bit of bisecting we found that this action became LocalPreferred after
D49580409, and from a brief look it seems that it wasn't intentional.

this diff changes it to the previous behaviour. Another option would be to add
`allow_cache_upload = True` for every CAF target, but it seems like a bigger
change.

Reviewed By: zsol

Differential Revision: D50168330

fbshipit-source-id: b83420dbc59d67ee85ca08b22c5aa2aee3c3892b