Updated Apple bundle info to take an app clip
Summary:
# Context

We are exploring the implementation of App Clips in Instagram to provide new ways to engage with the app and drive engagement. App Clips are a normal application that follows the normal application bundle structure and are shipped embedded in the main application bundle at the "AppClips/" folder.

Once uploaded to App Store Connect it's up to Apple how to accordingly vendor the IPA to make the App Clip version available to iOS without going through the store, while still sending the main app to the usual App Store CDN.

# What

App Clips live in the main app's bundle in the `AppClips` folder. This new destination path has been added in order to support the bundling of this bundle starting from the `ios_rule.bzl` definition, down to the `apple_bundle_part` logic responsible to copy it in the right place in the `.ipa`

Reviewed By: blackm00n

Differential Revision: D50661243

fbshipit-source-id: ae786998c65abb20f3b97611aec4ba2227ab0a51