Updated Apple Bundle extra attributes for an App Clip bundle
Summary:
# Context

We are exploring the implementation of App Clips in Instagram to provide new ways to engage with the app and drive engagement. App Clips are a normal application that follows the normal application bundle structure and are shipped embedded in the main application bundle at the "AppClips/" folder.

Once uploaded to App Store Connect it's up to Apple how to accordingly vendor the IPA to make the App Clip version available to iOS without going through the store, while still sending the main app to the usual App Store CDN.

# What

I have added a new extra attributes for apple bundles for App Clip bundles. Specifically the new attributes include the `is_appclip`.

Reviewed By: blackm00n

Differential Revision: D50661242

fbshipit-source-id: cd98fb335af6cd5148a097cf10c59a0279de4b9f