Revert D50699502: Multisect successfully blamed "D50699502: [fbcode][re][java_test] suport remote_execution attribute" for test or build failures
Summary:
This diff is reverting D50699502
D50699502: [fbcode][re][java_test] suport remote_execution attribute by snufkinsnorka has been identified to be causing the following test or build failures:

Tests affected:
- [multifeed/adsintegration/dependencies:run_verify_restrictions - main](https://www.internalfb.com/intern/test/844424991226658/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3418952
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D50765672

fbshipit-source-id: f2d80384e8864b38d43e39ae92e49cd40dcfd929