buck2 | android | Propagate the native library merging as a constraint value to platform when building android rules.
Summary:
Propagating this as constraint value, so we can apply select() in the dependencies/etc across the stack.

This is importan to be able to exclude linker flags in OpenXR and PanelAppSDK, since otherwise - the linking just starts to fail altogether.

somerge in buck2 is linking in a different way that isn't quite the identical behavior as buck1, and we want to move the codebase to not have to assume those behaviors, hence the select/config/constraint.

Reviewed By: cjhopman

Differential Revision: D50715684

fbshipit-source-id: 01ed2c9b8346e121c8159e3f9098f6772ebfc144