Add an option to prevent sh_binary from adding an extension
Summary:
`sh_binary` currently always appends a script filename extension (e.g.
`.sh` or `.bat`) to the provided script if it doesn't already have one.
This might not always be desirable, e.g. if your `main` is pointing to a
binary build target which already has a `.exe` extension, Windows will
fail to run it if you add the `.bat` extension. Add an option to let the
user control whether the extension is added.

Reviewed By: IanChilds

Differential Revision: D50997954

fbshipit-source-id: 7f19998d0a8abb64a804225a4e14d9cb01e508a3