dedupe apple resources when doing universal split transition
Summary:
There is a potential problem of duplicating a resource when `apple_bundle` contains a resource wrapping a universal binary which contains generated resource:

```
   bundle ----- resource ----- binary
         \                   /      /
          \                 /      /
       fat \               /arm64 /x86_64
            \             /      /
             \           /      /
              \         /      /
               \       /      /
                \     /      /
                 \   /      /
                   resource --- genrule
```

resource would be then included in bundle 2 times (once for fat platform via bundle, once via aux binary resources).

To avoid that introduce an automatic transition for `apple_resource` rule which will filter cpu and universal constraints (only when universal is on). That would make a resource a single node in a graph.

Reviewed By: milend

Differential Revision: D51156753

fbshipit-source-id: b8fe82c708daf0c3e03a0c7d72d6319e7562476e