Generalize signature of Rust binary rule impls
Summary:
These signatures are overspecified with the precise Provider types, unlike any other rule impl in the prelude. This makes it annoying to add new providers to existing rules, notably `RuntimeFilesInfo` in {D51222272}.

The important thing is just that a rule impl returns a list of various `Provider`s. `buck2 audit providers` is the intended way to debug what those providers are.

Reviewed By: diliop

Differential Revision: D51222269

fbshipit-source-id: 6b23c520a6813539debf68d226847974825ca278