Using idb-targets for simulator management
Summary:
We separating parts of `idb` to separate binaries. The first part we separated is whole simulator management (list, boot, create, etc.)
That is a responsibility of new `idb-targets` tool.
Benefits:
* We depend on user environment less
* Easier bug fixes and rollouts
* Better error propagation and safer implementation

The only thing left in idb that buck2 depend on is actual test run process that we have not separated yet.

Reviewed By: blackm00n

Differential Revision: D51395236

fbshipit-source-id: 76336c56facc24f1290611d06c36408347b86f31