Implement allowlisted_locales
Summary:
This was implemented in buck1 in D41142671. Unfortunately, that diff didn't bother to implement the buck2 version but just hardcoded it to an empty list. After tracking down some differences in buck1-built APK and buck2-built APKs, I found that this was the source of the difference :(

buck2 implementation just passes the `allowlisted_locales` through and the existing code takes care of it just the same.

Reviewed By: jiawei-lyu

Differential Revision: D51429439

fbshipit-source-id: ee63b3839344d0fe95badcbbf5408edb4a55c738