prelude: Different `-Cmetadata` for doctests
Summary:
`buck2 test //buck2/starlark-rust/starlark:starlark[doc]` fails with errors like this right now:

```rust
error[E0519]: the current crate is indistinguishable from one of its dependencies: it has the same crate-name `starlark` and was compiled with the same `-C metadata` arguments. This will result in symbol conflicts between the two.
   --> buck-out/v2/gen/fbcode/ca3f1b68f56b9220/buck2/starlark-rust/starlark/__starlark__/__srcs/src/values/typing/type_compiled/compiled.rs:157:1
    |
157 | / impl<'v, T: 'static> StarlarkValue<'v> for TypeCompiledImplAsStarlarkValue<T>
158 | | where
159 | |     T: TypeMatcher,
160 | | {
...   |
196 | |     }
197 | | }
    | |_^
```

This is a bit unsurprising - doctests depend on the non-doctest version of their crate, and we do currently compile both with the same metadata and crate name. What's is surprising is why this isn't causing more problems already. I guess rustc is just inconsistent about detecting this?

In any case, using a different `-Cmetadata` for doctests is pretty clearly the right thing to do. It's also what cargo does.

Reviewed By: zertosh

Differential Revision: D51580695

fbshipit-source-id: afa754a2ce6cde7dc8931973e2d79b230379acc8