Factor artifact outputs into a record
Summary:
Throughout the code dealing with resources, there is an assumption that there are only 3 relevant things to know about a resource:

1. A **name**, which determines the key that you will use when querying for the resource's location at runtime.

2. The **default output** of the target, which is the artifact whose filepath you will get back when you query the resource's location at runtime.

3. **Other outputs**, which are guaranteed to be materialized along with the resource.

Split DWARF does not fit into this formulation. Consider this scenario:

```
rust_binary(
    name = "main",
    srcs = [":gen-src"],
    resources = {"res": ":thing"},
)

genrule(
    name = "gen-src",
    out = "main.rs",
    cmd = "$(exe :code-generator) > $OUT",
)

rust_binary(
    name = "code-generator",
    srcs = ...,
    resources = {"res": ":thing"},
)

rust_binary(
    name = "thing",
    srcs = ...,
    deps = [
        ":thing-lib",
    ],
)

rust_library(
    name = "thing-lib",
    srcs = ...,
)
```

Suppose we build `buck2 build :main` with remote execution turned on, Split DWARF turned on, and in a build mode that uses dynamic linking in the execution platform. The executable produced by `:thing` is used in 2 ways:

- First, `:thing` needs to be available during the `:gen-src` action when `:code-generator` is executed. This runs in RE and there is no debugger involved. The shared library dependency `thing-lib.so` must be materialized, but external debuginfo for `:thing` should *not* be materialized during this action.

- Second, `:thing` needs to be materialized on the host so that when the build is done, and the user runs `:main` in a debugger, and `:main` launches `:thing` at runtime, it's available. **For this, we do need debuginfo** for `:thing`, in addition to the shared library deps.

From this it's clear that {name, default output, other outputs} is not enough information about a resource to execute this build as described. `thing.dwo` vs `thing-lib.so` are not both materialized in the same situations, so conflating them into the resource's `other_outputs` is not going to work.

Starting in D51222272, I am wiring up a way to distinguish runtime files that are always required whenever running an executable, vs runtime files that are only needed in a context where a user might be running a debugger.

This diff prepares for that by replacing 2-tuples `(Artifact, list[ArgLike])` which have been used so far to describe a resource's default output and other outputs, with an `ArtifactOutputs` record which will be able to convey debug-only other artifacts for a resource.

Reviewed By: zertosh

Differential Revision: D51737131

fbshipit-source-id: 457c2cac2b4a73678e93077d5a382f941d5a8fb3