Populate mandatory runtime files based on DistInfo provider
Summary:
Long explanation: see the summary of {D51737131}.

This diff does not immediately induce any change in behavior because, for now, every target that has a `DistInfo` provider is still putting identical contents into `DistInfo.nondebug_runtime_files` as into `DefaultInfo.other_outputs` (see D51222272). But later diffs in this stack begin to diverge the contents appropriately.

The change in linking/strip.bzl is to break an import cycle arising from the new dependency on dist_info.bzl. It's not great but I did not find any better way to resolve this one.

Reviewed By: zertosh

Differential Revision: D51742571

fbshipit-source-id: c94b998cfe42c8bf59a92da802cc0253793bebb3