Materialize external debuginfo any time cxx executable is a build output
Summary:
This commented-out line was added by D32588483, blocked on distinguishing runtime resources (needed for running an executable) vs external debuginfo (needed for debugging an executable) better throughout the prelude's rules.

https://www.internalfb.com/code/fbsource/[46737e406ac4e5f350e9bf9b2c96b64a2e75b6f4]/fbcode/buck2/prelude/cxx/cxx_executable.bzl?lines=573-576%2C578

The lower part of this stack has implemented the necessary improvements, so this is ready to enable now.

Previously, when you built a C++ binary as `buck2 build :main` with Split DWARF enabled, you would be guaranteed to materialize the binary itself, but it would not immediately be usable in a debugger. You would need to additionally build `:main[debuginfo]` to materialize unpacked external debuginfo (.dwo) or `:main[dwp]` to materialize packed external debuginfo (.dwp).

This diff makes `:main` automatically materialize all the unpacked external debuginfo referenced by that executable, making it immediately usable in a debugger.

Reviewed By: andrewjcg, christycylee

Differential Revision: D51189971

fbshipit-source-id: ab109486dc73f26167537e88de8c5d4c3fbc4d99