Add used_as_dependency_deprecated_do_not_use option to robolectric_test
Summary:
We don't want `robolectric_test`s to depend upon each other, it causes issues.

So, we'll not expose a `JavaLibraryInfo` to compile against unless `used_as_dependency_deprecated_do_not_use` is true.

This is buck2-only, which is fine as all (but 2) tests run with buck2 on CI.

Reviewed By: navidqar

Differential Revision: D52033063

fbshipit-source-id: 556ab7e8e2af6571a1c6d0da88191f2f64c8b16a