Add toolchain_alias as a rule (attempt 2)
Summary:
If you want to do anything cross compilation or anything, you pretty much have to introduce a toolchain alias. If Buck1 had toolchains, this would definitely be part of the prelude, and available by default. I appreciate we don't want to add new things to the prelude, but given how fundamental this is, I think we should do it.

Originally landed in D51535366, but due to a bug in the land the wrong version landed. Reverted that and trying again.

Reviewed By: JakobDegen

Differential Revision: D52045136

fbshipit-source-id: 9bf28cee15c10698a600139344e6f742eeb0fd38