add unzipped root directory to os.environ["PATH"] for windows
Summary:
On windows, `ctypes.util.find_library` searches the `PATH` environment variable for the requested dynamic library rather than the search path modified by `os.add_dll_directory`. Our pex packaging places dependent dynamic libraries in the root directory of the unzipped pex file, so that directory needs to be included in the PATH for `ctypes` to be able to find the dlls.

This impacts the ability of `usb.core` to find the `libusb-1.0.dll` backend, see example in test plan below.

Reviewed By: zsol

Differential Revision: D48687626

fbshipit-source-id: 9f33b9fb7a9b984fa504d84e3c92b49d6d3e2f4f