rules: `LibOutputStyle` in `lang_style_params`
Summary:
This takes the conversion from `LinkStrategy` to `LibOutputStyle` and moves it to where the providers are created, instead of doing it when building the params. This resolves the todo from Chris. Some weirdness is left behind, but that seems to mostly be fundamental, not incidental. We'll resolve that and clean some things up from here on forward

This diff is not a behavior change for most builds, but it does fix a bug with our subtargets, as demonstrated in the test

Reviewed By: dtolnay

Differential Revision: D52450310

fbshipit-source-id: 3f0ac6ab811e61a175b6356c84c7eb5c9d049ffd