rules: Bundle external debug info less aggressively
Summary:
The `external_debug_info` in the `LinkInfo` type is not intended to transitively include the debug info for all deps (this is despite it being a tset, I'm not entirely sure why that's the case, but as best I can tell it's something apple related. In any case, it works out for us). My evidence for that claim is that it's what C++ does/doesn't do. The full set of transitive debuginfo is instead stored in the `external_debug_info` field in the `MergedLinkInfo`, as well as some other places.

This diff updates the Rust rules to stop bundling the debuginfo from native deps in that field. We still include the debuginfo from Rust deps, for reasons that will be made clear next diff (tl;dr: staticlibs/cdylibs)

The `make_artifact_tset` call is not bundled into the new function for Next Diff reasons

Reviewed By: dtolnay

Differential Revision: D52499880

fbshipit-source-id: fd2e37bd6045b55216260b172aac6e5d0cfac3e4