Revert D52499870: Multisect successfully blamed "D52499870: [rust] rules: Fix pre-stripped artifacts" for otest failure
Summary:
This diff is reverting D52499870
D52499870: [rust] rules: Fix pre-stripped artifacts by JakobDegen has been identified to be causing the following test failure:

Tests affected:
- [tupperware/load_testing/scaletest/library:test_library_test - testRunCommand (tupperware.load_testing.scaletest.library.unittest.scaletest_util_test.ScaleTestUtilTests)](https://www.internalfb.com/intern/test/562950052913733/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3874616
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: artemy-kolesnikov

Differential Revision: D52539799

fbshipit-source-id: 0f0fcc639501f9907cfc1e9170df2c3f26ca35ac