Fix pool header block problem
Summary:
There seems to be a bug in the pool allocation regarding the header block in the scenario where no words are left (besides the header word).

Added a panic in the scenario I believe is broken and where the null terminating word is actually overwriting the header word.

In OCaml, any heap block has to be at least 2 words (1 header + one block word). The ocamlpool code needs to change so that when we cannot allocate the next block, there is one word left in the pool (in addition to the reserved header word).

That way, the zero terminator can be properly written in the String_tagged header block.

Reviewed By: rjbailey

Differential Revision: D52574962

fbshipit-source-id: 41795b3900551627e444cf627ad8642745a18434