rules: Drop native providers from `prebuilt_rust_library`
Summary: Without `advanced_unstable_linking`, it's not actually possible to correctly output native providers for prebuilt Rust libraries because we have no cdylib/staticlib build. With `advanced_unstable_linking`, it would be possible - however, the last diff removed what is afaik the only way in which that was ever used, and keeping this around would require extra work from me on the next few diffs. If someone wants to bring it back (or if it turns out that this is needed), we should refactor to share the code with the `rust_library` native provider creation

Reviewed By: capickett, dtolnay

Differential Revision: D52499875

fbshipit-source-id: 4648bbf0b510012aacf40faf62ccbb4beb975f42