halting instead of stopping the test binary
Summary:
We have experienced some issues where the ct_executor would take a long time to stop.

From that we have learnt the following lessons:
1) init:stop can take a long time to stop.
2) The code
```
 init:stop(ExitCode),
    receive
    after ?INIT_STOP_TIMEOUT ->
        ?LOG_ERROR(
            io_lib:format("~p failed to terminate within ~c millisecond", [
                ?MODULE, ?INIT_STOP_TIMEOUT
            ])
        ),
        erlang:halt(ExitCode)
    end;
```
That was supposed to prevent against that didn't do its job properly.

To prevent future problem like this one happening in the test binary, we do the same mitigation as with the ct_executor, that is, using halt/ instead of stop/

Reviewed By: TheGeorge

Differential Revision: D52877538

fbshipit-source-id: d5105a81d6d29f9395098629a53b87d9fd20e4de