Add easy install script for distro repo packages in BUCK files
Summary:
This script simplifies the process for installing repo packages when building
via buck2.

When developers run `buck2 run //build/buck2/install_deps`, the local
distribution / package manager will be detected, and the necessary packages
will be installed, based on tags attached to third party dependencies.

With the initial implementation, only fedora, ubuntu, and brew, are supported,
with additional environments / package managers to be considered in the future.

Reviewed By: bigfootjon

Differential Revision: D63086131

fbshipit-source-id: d0ab859b504615aa39c6ec54074daf365deb2fef