Set up .buckconfig / PACKAGE / cfg modifier / distro constraints
Summary:
This sets up the shims as its own buck2 project, with its own
.buckconfig, etc. This has some advantages in simplifying build/testing this
shim folder, as the target platform doesn't always need to be specified on the
CLI.

This diff also sets up a PACKAGE file, and sets up the cfg modifier init code
within them. This will allow us to set constraints on the CLI and
programatically enable/disable optional features of projects.

Reviewed By: bigfootjon

Differential Revision: D63653577

fbshipit-source-id: 3c7793800e794281014a971589a9f6060ecd87a0