Switch to bundled buck2 prelude
Summary: Using a git submodule for the buck2 prelude is not the way to go anymore. Earlier this year JakobDegen implemented a way to embed the prelude's contents into the buck executable. The bundled prelude is guaranteed to come from the same commit that the buck executable was built from, so there are not incompatibilities like in D64119241 (*Variable 'set' not found, did you mean 'int'?*).

Reviewed By: zertosh

Differential Revision: D64460123

fbshipit-source-id: 64fd2ed48f8128ee5d2606bc7b8971bbfe722d86