recover oss buck2 build
Summary:
this should land before configerator D67957414

fbsource D67939266 made a start on recovering builds by adding an empty file shim/third-party/ocaml/opam. this was sufficient to unblock the ocaml-setup.sh step in the github workflows and allowed us to recover the cargo build as a consequence. however when investigating how to recover the buck2 builds i've come to believe this was the wrong thing to do.

see configerator  D67957414 which changes the config to use shipit to copy in the shim directory restoring shim/third-party/ocaml/opam. for this to work, this diff removes shim/third-party/ocaml/opam from fbcode. in order to land this diff, both the cargo and buck2 tests are disabled.

when configerator D67957414 lands the cargo tests can be reenabled and hopefully we'll be able to reenable the reindeer step & buck2 tests too (see the changes to .buckconfig in this diff - they ought to help and certainly can't make anything worse).

Reviewed By: bigfootjon

Differential Revision: D67958197

fbshipit-source-id: b6a02a8f5573097e26e858dd57321f5b1a15231e