Remove and ignore .rustc_info.json files
Summary: {D73453607} still makes sense for the purpose of using a consistent rustc version to create the autocargo-generated Cargo.lock files, but the .rustc_info.json files should not be checked in. This avoids churn when different people run autocargo from an x86_64-unknown-linux-gnu host vs an aarch64-apple-darwin host or anything else.

Reviewed By: capickett

Differential Revision: D73455472

fbshipit-source-id: 4b04e0a6cdb04fee37cac07732e2751d693d1e04