Remove specialty handling of atomic lib in github shims
Summary:
X-link: https://github.com/facebook/buck2-shims-meta/pull/15

These lines were added a year ago in D62590238 but are now the apparent cause of an error in the buck2 OSS build. We could try to work around this on the buck side, but the specialty handling of this one particular library is nonstandard anyway.

Reviewed By: bigfootjon

Differential Revision: D81166319

fbshipit-source-id: fad24f6a3a9337f055bb9f7a6296e651dd8f9dde